modules/kali: Clean up library linking
The libasound library is not used, and pthread is added via other means,
so it doesn't need to be declared explicitly.

The Kali libraries themselves should be findable assuming they are in a
location that ld-linux knows about, and in the trial kali package I have,
the libraries are in usr/lib anyway.