baratinoo: Further improve reactivity
Let the process loop run as long as necessary in one go unless we get a
stop request, in which case terminate as soon as possible.

Theoretically, this both lift some load off the synthesis thread by not
leaving the Baratinoo TTS loop until actually useful, and makes the TTS
loop return as soon as it it tried to output some signal.