baratinoo: Try and handle punctuation levels
This is problematic because Baratinoo itself can't do that.  This means
we need to tell it explicitly how to handle the punctuation characters,
which is far from trivial.  In the current implementation it has an
annoying bug that it removes the punctuation properties of the
character, meaning that it won't alter intonation or introduce pauses
anymore.